feat(ops): ops host

+++ i2f-springboot/i2f-springboot-ops-starter/src/main/java/i2f/springboot/ops/host/controller/HostOpsController.java
@@ -1,11 +1,12 @@
 package i2f.springboot.ops.host.controller;
 
+import i2f.os.OsUtil;
+import i2f.springboot.ops.common.OpsException;
 import i2f.springboot.ops.common.OpsSecureDto;
 import i2f.springboot.ops.common.OpsSecureReturn;
 import i2f.springboot.ops.common.OpsSecureTransfer;
 import i2f.springboot.ops.host.data.HostFileItemDto;
 import i2f.springboot.ops.host.data.HostOperateDto;
-import i2f.springboot.ops.redis.data.RedisOperateDto;
 import i2f.web.servlet.ServletFileUtil;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -17,13 +18,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletResponse;
-import java.io.File;
-import java.io.PrintWriter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+import java.io.*;
+import java.security.MessageDigest;
+import java.util.*;
 
 /**
  * @author Ice2Faith
@@ -43,7 +43,7 @@ public class HostOpsController {
     @ResponseBody
     public OpsSecureReturn<OpsSecureDto> workdir(@RequestBody OpsSecureDto reqDto) throws Exception {
         try {
-            HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
+            Long req = transfer.recv(reqDto, Long.class);
             HostOperateDto resp=new HostOperateDto();
             File file = new File(".");
             file=new File(file.getAbsolutePath());
@@ -63,6 +63,8 @@ public class HostOpsController {
             String workdir = req.getWorkdir();
             File dir = new File(workdir);
             List<HostFileItemDto> resp=new ArrayList<>();
+            List<HostFileItemDto> dirList=new ArrayList<>();
+            List<HostFileItemDto> fileList=new ArrayList<>();
             if(dir.exists()){
                 File[] files = dir.listFiles();
                 if(files!=null){
@@ -73,24 +75,18 @@ public class HostOpsController {
                         item.setSize(file.length());
                         if(file.isDirectory()){
                             item.setType("dir");
+                            dirList.add(item);
                         }else{
                             item.setType("file");
+                            fileList.add(item);
                         }
-                        resp.add(item);
                     }
                 }
             }
 
-            resp.sort((f1,f2)->f1.getName().toLowerCase().compareTo(f2.getName().toLowerCase()));
+            dirList.sort((f1,f2)->f1.getName().toLowerCase().compareTo(f2.getName().toLowerCase()));
+            fileList.sort((f1,f2)->f1.getName().toLowerCase().compareTo(f2.getName().toLowerCase()));
 
-            if(dir!=null){
-                HostFileItemDto item=new HostFileItemDto();
-                item.setName(".");
-                item.setPath(new File(dir.getAbsolutePath()).getAbsolutePath());
-                item.setSize(0);
-                item.setType("dir");
-                resp.add(0,item);
-            }
             if(dir!=null){
                 HostFileItemDto item=new HostFileItemDto();
                 item.setName("..");
@@ -98,7 +94,23 @@ public class HostOpsController {
                 item.setPath(parentFile!=null?parentFile.getAbsolutePath():"/");
                 item.setSize(0);
                 item.setType("dir");
-                resp.add(0,item);
+                resp.add(item);
+            }
+
+            if(dir!=null){
+                HostFileItemDto item=new HostFileItemDto();
+                item.setName(".");
+                item.setPath(new File(dir.getAbsolutePath()).getAbsolutePath());
+                item.setSize(0);
+                item.setType("dir");
+                resp.add(item);
+            }
+
+            resp.addAll(dirList);
+            resp.addAll(fileList);
+
+            for (HostFileItemDto item : resp) {
+                item.setSizeText(humanFileSize(item.getSize()));
             }
 
             return transfer.success(resp);
@@ -108,6 +120,22 @@ public class HostOpsController {
         }
     }
 
+    public static String humanFileSize(long size){
+        if(size<1024){
+            return String.format("%dB", size);
+        }
+        if(size<1024*1024){
+            return String.format("%.2fKB", size / 1024.0);
+        }
+        if(size<1024*1024*1024){
+            return String.format("%.2fMB", size / 1024.0 / 1024.0);
+        }
+        if(size<1024L*1024*1024*1024){
+            return String.format("%.2fGB", size / 1024.0 / 1024.0 / 1024.0);
+        }
+        return String.format("%.2fTB", size / 1024.0 / 1024.0 / 1024.0 / 1024.0);
+    }
+
     @PostMapping("/delete")
     @ResponseBody
     public OpsSecureReturn<OpsSecureDto> delete(@RequestBody OpsSecureDto reqDto) throws Exception {
@@ -115,12 +143,27 @@ public class HostOpsController {
             HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
             String path = req.getPath();
             File file = new File(path);
-            boolean resp = file.delete();
-            return transfer.success(resp);
+            deleteFile(file);
+            return transfer.success(true);
         } catch (Throwable e) {
             log.warn(e.getMessage(),e);
             return transfer.error(e.getMessage());
         }
+    }
+
+    public void deleteFile(File file) throws Exception {
+        if(file==null){
+            return;
+        }
+        if(file.isDirectory()){
+            File[] files = file.listFiles();
+            if(files!=null){
+                for (File item : files) {
+                    deleteFile(item);
+                }
+            }
+        }
+        file.delete();
     }
 
     @PostMapping("/mkdirs")
@@ -155,4 +198,179 @@ public class HostOpsController {
         }
     }
 
+    @PostMapping("/tail")
+    public void tail(@RequestBody OpsSecureDto reqDto, HttpServletResponse response) throws Exception {
+        try {
+            HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
+            String path = req.getPath();
+            int lineCount = req.getLineCount();
+            if(lineCount<=0){
+                lineCount=100;
+            }
+            File file = new File(path);
+            int count=0;
+            LinkedList<String> lines=new LinkedList<>();
+            try(BufferedReader reader = new BufferedReader(new FileReader(file))) {
+                String line = null;
+                while ((line = reader.readLine()) != null) {
+                    lines.add(line);
+                    count++;
+                    if (count > lineCount) {
+                        lines.removeFirst();
+                        count--;
+                    }
+                }
+            }
+            StringBuilder builder=new StringBuilder();
+            for (String item : lines) {
+                builder.append(item).append("\n");
+            }
+            ByteArrayInputStream bis = new ByteArrayInputStream(builder.toString().getBytes());
+            ServletFileUtil.responseAsFileAttachment(bis,true,file.getName(),null,!req.isInline(),response);
+        } catch (Throwable e) {
+            log.warn(e.getMessage(),e);
+            response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
+            PrintWriter writer = response.getWriter();
+            writer.write("Internal Server Error");
+            writer.flush();
+            return;
+        }
+    }
+
+
+    @PostMapping("/head")
+    public void head(@RequestBody OpsSecureDto reqDto, HttpServletResponse response) throws Exception {
+        try {
+            HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
+            String path = req.getPath();
+            int lineCount = req.getLineCount();
+            if(lineCount<=0){
+                lineCount=100;
+            }
+            File file = new File(path);
+            int count=0;
+            StringBuilder builder=new StringBuilder();
+            try(BufferedReader reader = new BufferedReader(new FileReader(file))) {
+                String line = null;
+                while ((line = reader.readLine()) != null) {
+                    builder.append(line).append("\n");
+                    count++;
+                    if (count > lineCount) {
+                        break;
+                    }
+                }
+            }
+
+            ByteArrayInputStream bis = new ByteArrayInputStream(builder.toString().getBytes());
+            ServletFileUtil.responseAsFileAttachment(bis,true,file.getName(),null,!req.isInline(),response);
+        } catch (Throwable e) {
+            log.warn(e.getMessage(),e);
+            response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
+            PrintWriter writer = response.getWriter();
+            writer.write("Internal Server Error");
+            writer.flush();
+            return;
+        }
+    }
+
+    @PostMapping("/upload")
+    @ResponseBody
+    public OpsSecureReturn<OpsSecureDto> upload(MultipartFile file, OpsSecureDto reqDto) throws Exception {
+        try {
+            HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
+            String path = req.getPath();
+            File saveFile=null;
+            if(path!=null && !path.isEmpty()){
+                saveFile = new File(path);
+            }else {
+                String workdir = req.getWorkdir();
+                File dir = new File(workdir);
+                String originalFilename = file.getOriginalFilename();
+                if (originalFilename == null || originalFilename.isEmpty()) {
+                    throw new OpsException("missing origin file name");
+                }
+                File originFile = new File(originalFilename);
+                saveFile = new File(dir, originFile.getName());
+            }
+            File tmpFile=new File(saveFile.getParentFile(),"upload-"+ (UUID.randomUUID().toString().replace("-",""))+".tmp");
+            try {
+                MessageDigest digest = MessageDigest.getInstance("MD5");
+                OutputStream os = new FileOutputStream(tmpFile);
+                byte[] buffer = new byte[2048];
+                int len = 0;
+                InputStream is = file.getInputStream();
+                while ((len = is.read(buffer)) > 0) {
+                    digest.update(buffer, 0, len);
+                    os.write(buffer, 0, len);
+                }
+                os.close();
+                byte[] bytes = digest.digest();
+                StringBuilder builder = new StringBuilder();
+                for (byte bt : bytes) {
+                    builder.append(String.format("%02x", (int) (bt & 0x0ff)));
+                }
+                String calcMd5 = builder.toString();
+                if (!calcMd5.equalsIgnoreCase(req.getMd5())) {
+                    throw new OpsException("file check sum error");
+                }
+                tmpFile.renameTo(saveFile);
+                return transfer.success(true);
+            }finally {
+                if(tmpFile!=null && tmpFile.exists()){
+                    tmpFile.delete();
+                }
+            }
+        } catch (Throwable e) {
+            log.warn(e.getMessage(),e);
+            return transfer.error(e.getMessage());
+        }
+    }
+
+
+    @PostMapping("/cmd")
+    @ResponseBody
+    public OpsSecureReturn<OpsSecureDto> cmd(@RequestBody OpsSecureDto reqDto) throws Exception {
+        try {
+            HostOperateDto req = transfer.recv(reqDto, HostOperateDto.class);
+            String workdir = req.getWorkdir();
+            String cmd = req.getCmd();
+            boolean runAsFile = req.isRunAsFile();
+            long waitForSeconds = req.getWaitForSeconds();
+            if(waitForSeconds<0){
+                waitForSeconds=120;
+            }
+            if(waitForSeconds>=500){
+                waitForSeconds=500;
+            }
+            File bashFile=null;
+            try {
+                String resp =null;
+                if(runAsFile){
+                    String suffix=".sh";
+                    if(OsUtil.isWindows()){
+                        suffix=".bat";
+                    }
+                    bashFile = new File(workdir,"cmd-"+(UUID.randomUUID().toString().replace("-",""))+suffix);
+                    try(FileOutputStream fos = new FileOutputStream(bashFile)){
+                        fos.write(cmd.getBytes(OsUtil.getCmdCharset()));
+                    }
+                    String[] cmdArr=new String[]{"sh","-c",bashFile.getName()};
+                    if(OsUtil.isWindows()){
+                        cmdArr=new String[]{"cmd.exe","/c",bashFile.getName()};
+                    }
+                    resp=OsUtil.execCmd(true, waitForSeconds, cmdArr, null, new File(workdir), null);
+                }else {
+                    resp=OsUtil.execCmd(true, waitForSeconds, cmd, null, new File(workdir), null);
+                }
+                return transfer.success(resp);
+            }finally {
+                if(bashFile!=null && bashFile.exists()){
+                    bashFile.delete();
+                }
+            }
+        } catch (Throwable e) {
+            log.warn(e.getMessage(),e);
+            return transfer.error(e.getMessage());
+        }
+    }
 }
